backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
numpy==1.26.4
scipy==1.13.1
reportlab==4.2.0
stripe==10.4.0
python-dotenv==1.0.1
slowapi==0.1.9
mailchimp-marketing==3.0.80
cryptography==42.0.8
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
